VoidProcTransaction
Added feature allowing users to disable the default feature that wraps all custom sprocs that return void in a transaction.

By default this feature is left turned on.

One small piece of this feature ended up in the formatting commit in file ComponentDataAccess.cst lines 1527 and 1582, most of it is in this commit.