Extended property feature
For the MS-SQL data access layer, add the ability to create custom Extended properties in the database on database columns (not tables).  There is a setting allowing the custom property prefix to be customized.  These custom properties are then written to each column enum value and can be used in code.

This provides a way to specify additional meta information in the database as the column level.