Large pull request including 3 new features and three upgrades (#477)
* Formatting only

These changes are only formatting changes, no functionality changes.

Newlines, removing trailing spaces, and replacing tabs with spaces.

* VoidProcTransaction

Added feature allowing users to disable the default feature that wraps all custom sprocs that return void in a transaction.

By default this feature is left turned on.

One small piece of this feature ended up in the formatting commit in file ComponentDataAccess.cst lines 1527 and 1582, most of it is in this commit.

* ConnectionString in error

Add the connection string information without credentials to the custom error that gets thrown when a sql call blows up.

* Extended property feature

For the MS-SQL data access layer, add the ability to create custom Extended properties in the database on database columns (not tables).  There is a setting allowing the custom property prefix to be customized.  These custom properties are then written to each column enum value and can be used in code.

This provides a way to specify additional meta information in the database as the column level.

* Param sql string builder ' bug

Fixes a bug in the parameterized sql builder where tick marks get double encoded.  Also includes a couple other minor fixes to the sql string builder.

* New base classes

Until now, All table entities had a base class they inherrited from.
This changes gives all View entities their own base class, and Also creates a single overall base class for all netTiers table and view entities.

This allows methods to pass around objects at the view, table, or netTiers level.

* New Column attribute helper methods

New helper method used to get the custom attributes off of column enum entities.  Is an addition to the custom column attributes in sql feature.