Refresh views with just chosen
The ExecuteViews feature has a bug in it where after execution the view list is populated with all views from the database, not just the chosen views.  This causes code to be generated for all the views, not just the desired ones.

This commit fixes that bug, so only the chosen views remain chosen.